+++ aten/src/ATen/native/quantized/cpu/qnnpack_maxpool.cpp
@@ -1,11 +1,12 @@
-#include "init_qnnpack.h"
-
 #include <ATen/ATen.h>
 #include <ATen/core/Type.h>
 #include <ATen/core/op_registration/op_registration.h>
 #include <THNN/generic/pooling_shape.h>
 
-namespace at { namespace native {
+#include "init_qnnpack.h"
+
+namespace at {
+namespace native {
 namespace {
 
 class QNNPACKMaxPool final : public c10::OperatorKernel {
@@ -16,10 +17,10 @@ class QNNPACKMaxPool final : public c10::OperatorKernel {
       std::vector<int64_t> stride,
       std::vector<int64_t> dilation,
       std::vector<int64_t> padding) {
-
     Tensor qy;
 
-    TORCH_CHECK(input.ndimension() == 4,
+    TORCH_CHECK(
+        input.ndimension() == 4,
         "qnnpack_maxpool(): Expected input to be 4-dimensional: got ",
         input.ndimension());
 
@@ -27,8 +28,8 @@ class QNNPACKMaxPool final : public c10::OperatorKernel {
 
     initQNNPACK();
 
-    const auto scale = input_contig.q_scale().toDouble();
-    const auto zero_point = input_contig.q_zero_point().toInt();
+    const auto scale = input_contig.q_scale();
+    const auto zero_point = input_contig.q_zero_point();
     qnnp_operator_t qnnpackOperator_{nullptr};
 
     int64_t padL = padding[0];
@@ -41,7 +42,8 @@ class QNNPACKMaxPool final : public c10::OperatorKernel {
     int64_t dilationW = dilation[1];
 
     TORCH_CHECK(kH > 0 && kW > 0, "kernel_size should be greater than zero.");
-    TORCH_CHECK(strideH > 0 && strideW > 0, "strides should be greater than zero.");
+    TORCH_CHECK(
+        strideH > 0 && strideW > 0, "strides should be greater than zero.");
 
     // Input is in NHWC format
     int64_t batch_size = input_contig.size(0);
@@ -65,22 +67,25 @@ class QNNPACKMaxPool final : public c10::OperatorKernel {
         std::numeric_limits<uint8_t>::max() /* output max */,
         0 /* flags */,
         &qnnpackOperator_);
-    TORCH_INTERNAL_ASSERT(createStatus == qnnp_status_success,
+    TORCH_INTERNAL_ASSERT(
+        createStatus == qnnp_status_success,
         "failed to create QNNPACK MaxPool operator");
     TORCH_INTERNAL_ASSERT(qnnpackOperator_ != nullptr);
 
     int64_t outC = inC;
-    int64_t outH = pooling_output_shape(inH, kH, padT, strideH, dilationH, false);
-    int64_t outW = pooling_output_shape(inW, kW, padL, strideW, dilationW, false);
+    int64_t outH =
+        pooling_output_shape(inH, kH, padT, strideH, dilationH, false);
+    int64_t outW =
+        pooling_output_shape(inW, kW, padL, strideW, dilationW, false);
 
-    TORCH_CHECK(outH > 0 && outW > 0, "the resulting output Tensor size should be >= 0");
+    TORCH_CHECK(
+        outH > 0 && outW > 0,
+        "the resulting output Tensor size should be >= 0");
 
     // NHWC
     std::vector<int64_t> outSizes{batch_size, outH, outW, outC};
-    qy = at::_empty_affine_quantized(outSizes,
-                                     at::device(kCPU).dtype(kQUInt8),
-                                     scale,
-                                     zero_point);
+    qy = at::_empty_affine_quantized(
+        outSizes, at::device(kCPU).dtype(kQUInt8), scale, zero_point);
 
     const qnnp_status setupStatus = qnnp_setup_max_pooling2d_nhwc_u8(
         qnnpackOperator_ /* max pooling */,
@@ -93,21 +98,24 @@ class QNNPACKMaxPool final : public c10::OperatorKernel {
         outC /* output_pixel_stride */,
         nullptr /* thread pool */);
 
-    TORCH_INTERNAL_ASSERT(setupStatus == qnnp_status_success,
+    TORCH_INTERNAL_ASSERT(
+        setupStatus == qnnp_status_success,
         "failed to setup QNNPACK MaxPool operator");
 
     const qnnp_status runStatus =
         qnnp_run_operator(qnnpackOperator_, nullptr /* thread pool */);
 
-    TORCH_INTERNAL_ASSERT( runStatus == qnnp_status_success,
-       "failed to run QNNPACK MaxPool operator");
+    TORCH_INTERNAL_ASSERT(
+        runStatus == qnnp_status_success,
+        "failed to run QNNPACK MaxPool operator");
     return qy;
   }
 };
 
 static auto registry = c10::RegisterOperators().op(
     "quantized::qnnpack_max_pool2d(Tensor input, int[] kernel_size, int[] stride, int[] dilation, int[] padding) -> Tensor",
-    c10::RegisterOperators::options()
-      .kernel<QNNPACKMaxPool>(QuantizedCPUTensorId()));
-}  // namespace
-}}  // namespace at::native
+    c10::RegisterOperators::options().kernel<QNNPACKMaxPool>(
+        QuantizedCPUTensorId()));
+} // namespace
+} // namespace native
+} // namespace at
